1. Refactored method addCar 2. Created package for enum Role
