+++ src/main/java/innowise/khorsun/carorderservice/util/enums/Role.java
@@ -1,0 +1,15 @@
+package innowise.khorsun.carorderservice.util.enums;
+
+public enum Role {
+    MANAGER("MANAGER"),
+    CUSTOMER("CUSTOMER");
+    private final String val;
+
+    Role(String val) {
+        this.val = val;
+    }
+
+    public String getVal() {
+        return val;
+    }
+}
